requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=3.3.0

# Data Processing
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Progress & Utilities
tqdm>=4.62.0
pyyaml>=5.4.0

# Optional: Spotify API Integration
# spotipy>=2.19.0

# Optional: Web Interface
# streamlit>=1.10.0
# flask>=2.0.0

# Development
pytest>=6.2.0
black>=21.0
